added helpful log, removed util.rs in favour of env vars
Now the GLIB type coercion errors will be followed by a 'please ignore that stupid lib' message so you can know it isn't an actual message and the lib is just eagerly printing to stderr. And since util.rs is a major security issue (not for our users, but for us) since we might be prone to accidentally pushing our root password to GitHub, we switched to environment variables (which also means that the person presenting would need to export the SUDO_PASS env var before presenting so that app runs smoothly.